requirements.txt
pytest
pylint
ipdb
google-cloud-storage==2.1.0
gcsfs
jupyterlab
tensorflow==2.7.0
numpy
pandas
matplotlib
seaborn
scikit-learn
fastapi
uvicorn
python-multipart
retina-face
